# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Compass Hub API",
    "version": "v1",
    "description": "Read-only consolidated employee time and work-package API"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "tags": [
    {
      "name": "Consolidated View",
      "description": "Employee-centered consolidated read API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SourceSystemSnapshot": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded",
              "unavailable"
            ]
          },
          "lastSyncedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "status",
          "lastSyncedAt"
        ]
      },
      "TimeEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "sourceSystem": {
            "type": "string",
            "minLength": 1
          },
          "workPackageId": {
            "type": "string",
            "minLength": 1
          },
          "date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "hours": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "lastUpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sourceSystem",
          "workPackageId",
          "date",
          "hours",
          "lastUpdatedAt"
        ]
      },
      "WorkPackage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "sourceSystem": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "assigned",
              "active",
              "blocked",
              "closed"
            ]
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sourceSystem",
          "name",
          "status",
          "assignedAt",
          "lastUpdatedAt"
        ]
      },
      "ConsolidatedEmployeeView": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "minLength": 1
          },
          "asOf": {
            "type": "string",
            "format": "date-time"
          },
          "freshnessLagSeconds": {
            "type": "integer",
            "minimum": 0
          },
          "sourceSystems": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "healthy",
                    "degraded",
                    "unavailable"
                  ]
                },
                "lastSyncedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "name",
                "status",
                "lastSyncedAt"
              ]
            }
          },
          "timeEntries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "minLength": 1
                },
                "sourceSystem": {
                  "type": "string",
                  "minLength": 1
                },
                "workPackageId": {
                  "type": "string",
                  "minLength": 1
                },
                "date": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                },
                "hours": {
                  "type": "number",
                  "exclusiveMinimum": 0
                },
                "lastUpdatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "id",
                "sourceSystem",
                "workPackageId",
                "date",
                "hours",
                "lastUpdatedAt"
              ]
            }
          },
          "workPackages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "minLength": 1
                },
                "sourceSystem": {
                  "type": "string",
                  "minLength": 1
                },
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "assigned",
                    "active",
                    "blocked",
                    "closed"
                  ]
                },
                "assignedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "lastUpdatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "id",
                "sourceSystem",
                "name",
                "status",
                "assignedAt",
                "lastUpdatedAt"
              ]
            }
          }
        },
        "required": [
          "employeeId",
          "asOf",
          "freshnessLagSeconds",
          "sourceSystems",
          "timeEntries",
          "workPackages"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/api/v1/employees/{employeeId}/consolidated-view": {
      "get": {
        "operationId": "getEmployeeConsolidatedView",
        "summary": "Get consolidated time entries and work packages for an employee",
        "tags": [
          "Consolidated View"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "employeeId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Consolidated employee view",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "employeeId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "asOf": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "freshnessLagSeconds": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "sourceSystems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "healthy",
                              "degraded",
                              "unavailable"
                            ]
                          },
                          "lastSyncedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "name",
                          "status",
                          "lastSyncedAt"
                        ]
                      }
                    },
                    "timeEntries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "sourceSystem": {
                            "type": "string",
                            "minLength": 1
                          },
                          "workPackageId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "date": {
                            "type": "string",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                          },
                          "hours": {
                            "type": "number",
                            "exclusiveMinimum": 0
                          },
                          "lastUpdatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "sourceSystem",
                          "workPackageId",
                          "date",
                          "hours",
                          "lastUpdatedAt"
                        ]
                      }
                    },
                    "workPackages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "sourceSystem": {
                            "type": "string",
                            "minLength": 1
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "assigned",
                              "active",
                              "blocked",
                              "closed"
                            ]
                          },
                          "assignedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "lastUpdatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "sourceSystem",
                          "name",
                          "status",
                          "assignedAt",
                          "lastUpdatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "employeeId",
                    "asOf",
                    "freshnessLagSeconds",
                    "sourceSystems",
                    "timeEntries",
                    "workPackages"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not authorized for this employee",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Employee record not found in the consolidated view",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "webhooks": {}
}